# wsc-meso/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: dev@mesocycleplanner.com
    name: Mesocycle Planner Development Team
  description: |
    Complete API for fitness training app with advanced periodization support.
    Features: 440+ exercises database, Linear/DUP/Block/Polarized periodization models,
    microcycle management, auto-progression, session logging, and AI-generated training plans.

    Supports user authentication, mesocycle planning, workout tracking, and progress analytics.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Mesocycle Planner API - Advanced Periodization
  version: 1.0.0
servers:
- description: Local Development
  url: http://localhost:8000/api/v1
- description: Staging
  url: https://staging.mesocycleplanner.com/api/v1
- description: Production
  url: https://api.mesocycleplanner.com/api/v1
...
paths:
  /auth/register:
    post:
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: User created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Invalid input
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: User already exists
      summary: Register a new user
      tags:
      - Authentication
  /auth/login:
    post:
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
          description: Login successful
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Invalid credentials
      summary: Login user
      tags:
      - Authentication
...
components:
...
  schemas:
...
    TrainingLevel:
      enum:
      - beginner
      - intermediate
      - advanced
      - elite
      title: TrainingLevel
      type: string
...
    User:
      example:
        full_name: John Doe
        updated_at: 2000-01-23T04:56:07.000+00:00
        training_level: beginner
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 123e4567-e89b-12d3-a456-426614174000
        email: user@example.com
        username: athlete123
      properties:
        id:
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          title: id
          type: string
        email:
          example: user@example.com
          format: email
          title: email
          type: string
        username:
          example: athlete123
          title: username
          type: string
        full_name:
          example: John Doe
          title: full_name
          type: string
        training_level:
          $ref: "#/components/schemas/TrainingLevel"
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      required:
      - created_at
      - email
      - id
      - username
      title: User
    UserCreate:
      example:
        password: password
        full_name: full_name
        training_level: beginner
        email: email
        username: username
      properties:
        email:
          format: email
          title: email
          type: string
        username:
          minLength: 3
          title: username
          type: string
        password:
          minLength: 8
          title: password
          type: string
        full_name:
          title: full_name
          type: string
        training_level:
          $ref: "#/components/schemas/TrainingLevel"
      required:
      - email
      - password
      - username
      title: UserCreate
    UserLogin:
      example:
        password: password
        email: email
      properties:
        email:
          format: email
          title: email
          type: string
        password:
          title: password
          type: string
      required:
      - email
      - password
      title: UserLogin
    Token:
      example:
        access_token: access_token
        token_type: bearer
        expires_in: 3600
      properties:
        access_token:
          title: access_token
          type: string
        token_type:
          example: bearer
          title: token_type
          type: string
        expires_in:
          example: 3600
          title: expires_in
          type: integer
      title: Token
...
    Error:
      example:
        details: "{}"
        error: error
        message: message
      properties:
        error:
          title: error
          type: string
        message:
          title: message
          type: string
        details:
          title: details
          type: object
      title: Error
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http